refactor: move app_bloc files to auth folder
