+++ lib/presentation/app/bloc/auth/app_event.dart
