+++ lib/presentation/app/bloc/auth/app_state.dart
