+++ lib/presentation/shared/router/go_router.dart
@@ -8,8 +8,8 @@ import 'package:on_time_front/domain/entities/preparation_entity.dart';
 import 'package:on_time_front/domain/entities/schedule_entity.dart';
 import 'package:on_time_front/presentation/alarm/screens/alarm_screen.dart';
 import 'package:on_time_front/presentation/alarm/screens/schedule_start_screen.dart';
+import 'package:on_time_front/presentation/app/bloc/auth/app_bloc.dart';
 import 'package:on_time_front/presentation/early_late/screens/early_late_screen.dart';
-import 'package:on_time_front/presentation/app/bloc/app_bloc.dart';
 import 'package:on_time_front/presentation/calendar/screens/calendar_screen.dart';
 import 'package:on_time_front/presentation/home/screens/home_screen_tmp.dart';
 import 'package:on_time_front/presentation/login/screens/sign_in_main_screen.dart';
